Agent3- PDF Classifier_OCR/requirements.txt
flask
google-generativeai
python-dotenv
pytesseract
pdfplumber
python-docx
Pillow
watchdog
jinja2
werkzeug
